Code reorganisation
Traceroute has been moved to a model package, and now store the local
list of ip.
Traceroute now contains an inner class Tracerouter, that launch a new
traceroute and update the local ipList with it.

Other part of the code have been changed so it works with this new
organisation